@@ -1,5 +1,5 @@
 {
-  "name": "advanced-ng",
+  "name": "ng-gl42023",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
@@ -10,33 +10,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.0.0",
-    "@angular/common": "^15.0.0",
-    "@angular/compiler": "^15.0.0",
-    "@angular/core": "^15.0.0",
-    "@angular/forms": "^15.0.0",
-    "@angular/platform-browser": "^15.0.0",
-    "@angular/platform-browser-dynamic": "^15.0.0",
-    "@angular/router": "^15.0.0",
-    "@faker-js/faker": "^7.6.0",
-    "bootstrap": "^5.2.3",
-    "chart.js": "^4.0.1",
-    "plotly.js-dist-min": "^2.6.2",
-    "rxjs": "~7.5.0",
+    "@angular/animations": "^16.1.0",
+    "@angular/common": "^16.1.0",
+    "@angular/compiler": "^16.1.0",
+    "@angular/core": "^16.1.0",
+    "@angular/forms": "^16.1.0",
+    "@angular/platform-browser": "^16.1.0",
+    "@angular/platform-browser-dynamic": "^16.1.0",
+    "@angular/router": "^16.1.0",
+    "@faker-js/faker": "^8.3.1",
+    "bootstrap": "^5.3.2",
+    "chart.js": "^4.4.1",
+    "jquery": "^3.7.1",
+    "memo-decorator": "^2.0.1",
+    "ngx-toastr": "^16.2.0",
+    "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.12.0"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.0.1",
-    "@angular/cli": "~15.0.1",
-    "@angular/compiler-cli": "^15.0.0",
+    "@angular-devkit/build-angular": "^16.1.5",
+    "@angular/cli": "~16.1.5",
+    "@angular/compiler-cli": "^16.1.0",
     "@types/jasmine": "~4.3.0",
-    "jasmine-core": "~4.5.0",
+    "jasmine-core": "~4.6.0",
     "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.0",
+    "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~4.8.2"
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~5.1.3"
   }
 }
