@@ -1,43 +1,53 @@
-import {Component, OnInit} from '@angular/core';
-import {User, UsersService} from "../users.service";
+import { Component, NgZone, OnInit } from '@angular/core';
+import { User, UsersService } from '../users.service';
 import * as ChartJs from 'chart.js/auto';
 @Component({
   selector: 'app-rh',
   templateUrl: './rh.component.html',
-  styleUrls: ['./rh.component.css']
+  styleUrls: ['./rh.component.css'],
 })
 export class RhComponent implements OnInit {
   oddUsers: User[];
   evenUsers: User[];
   chart: any;
-  constructor(private userService: UsersService) {
+  constructor(private userService: UsersService, private ngZone: NgZone) {
     this.oddUsers = this.userService.getOddOrEven(true);
     this.evenUsers = this.userService.getOddOrEven();
   }
 
   ngOnInit(): void {
-        this.createChart();
-    }
-  addUser(list: User[], newUser: string) {
-    this.userService.addUser(list, newUser);
+    this.loadChart();
   }
-  createChart(){
+
+  loadChart() {
+    this.ngZone.runOutsideAngular(() => {
+      this.createChart();
+    });
+  }
+  addUser(list: User[], newUser: string): void {
+    setTimeout(() => {
+      list = this.userService.addUser(list, newUser);
+      this.chart.destroy();
+      this.loadChart();
+    });
+  }
+
+  createChart() {
     const data = [
       { users: 'Workers', count: this.oddUsers.length },
       { users: 'Boss', count: this.evenUsers.length },
     ];
-    this.chart = new ChartJs.Chart("MyChart",
-    {
+    this.chart = new ChartJs.Chart('MyChart', {
       type: 'bar',
-        data: {
-          labels: data.map(row => row.users),
+      data: {
+        labels: data.map((row) => row.users),
         datasets: [
-        {
-          label: 'Entreprise stats',
-          data: data.map(row => row.count)
-        }
-      ]
-    }
+          {
+            label: 'Entreprise stats',
+            data: data.map((row) => row.count),
+          },
+        ],
+      },
     });
   }
 }
